% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDist.R
\name{calcDist}
\alias{calcDist}
\title{Calculate distribution of patterns in data.}
\usage{
calcDist(i, x, freq, depth, seqVec, nmin = 1, lik = F)
}
\arguments{
\item{i}{Index, to choose right period.}

\item{x}{Data, 3d array}

\item{freq}{Length of the periods}

\item{depth}{depth argument supplied to pstree-function}

\item{seqVec}{Vector of sequences to look for}

\item{nmin}{Argument passed to pstree}

\item{lik}{Argument passed to pstree}
}
\description{
This function calculates the distribution of patterns given in seqVec in the data, made to be used with (mc)lapply in findDist.
}
\examples{
calcDist()
}
\keyword{calc,}
\keyword{calculate}
\keyword{dist,}
\keyword{distribution,}
\keyword{pattern,}
\keyword{patterns,}
