% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIQDparallel.R
\name{calculateIQDparallel}
\alias{calculateIQDparallel}
\title{Calculate the IQD in parallel}
\usage{
calculateIQDparallel(
  i,
  RCM_input_list,
  dataset_input_list,
  output_list,
  seasons_list,
  remove_indices = NULL
)
}
\arguments{
\item{i}{input index for parallization, specifies what portion of data to work on}

\item{RCM_input_list}{List of paths to input climamodel data}

\item{dataset_input_list}{List of paths to input dataproduct data}

\item{output_list}{List of output paths}

\item{remove_indices}{Which indices to remove. Defaults to NULL}

\item{season_list}{List of seasons to calculate for}
}
\description{
This function calculates the IQD between data stored at paths at index i in the two input lists.
}
\examples{
calculateIQDparallel()
}
\keyword{IQD,}
\keyword{parallel}
