% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDiff.R
\name{findDiff}
\alias{findDiff}
\title{Find difference between two datasets}
\usage{
findDiff(a, b, relative = F)
}
\arguments{
\item{a}{Data, 3d array}

\item{b}{Data, 3d array}

\item{relative}{If true, returns relative difference.}
}
\description{
This function finds the difference between the mean for each grid in a and b (a-b)
}
\examples{
findDiff()
}
\keyword{dataHandling}
\keyword{diff,}
\keyword{difference,}
