% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDist.R
\name{findDist}
\alias{findDist}
\title{Finds distribution of patterns in data.}
\usage{
findDist(data, depth, seqVec, threshold = 0.1)
}
\arguments{
\item{data}{Data, 3d array}

\item{depth}{depth argument supplied to pstree-function}

\item{seqVec}{Vector of sequences to look for}

\item{threshold}{Threshold for seting values to zero. Defaults to 0.1.}
}
\description{
This function findes the distribution of patterns given in seqVec in the data.
}
\examples{
findDist()
}
\keyword{dist,}
\keyword{distribution}
\keyword{pattern,}
\keyword{patterns,}
