% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDistBin.R
\name{findDistBin}
\alias{findDistBin}
\title{Finds distribution of binomial pattern in data.}
\usage{
findDistBin(data, threshold = 0.1)
}
\arguments{
\item{data}{Data, 3d array}

\item{threshold}{Threshold for seting values to zero. Defaults to 0.1.}
}
\description{
This function sets values bellow threshold to zero, and find distribution of 1's and 0's in given data.
}
\examples{
findDistBin()
}
\keyword{dist,}
\keyword{distribution}
\keyword{pattern,}
\keyword{patterns,}
