% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFreq.R
\name{findFreq}
\alias{findFreq}
\title{Finds frequency of wet days}
\usage{
findFreq(data, threshold = 0.1)
}
\arguments{
\item{data}{Data, 3d array}

\item{treshold}{Datapoints with values less then treshold are forced to zero.}
}
\description{
This function returns a matrix with frequencey of wet days per gridpoint.
}
\examples{
findFreq()
}
\keyword{dataHandling}
\keyword{freq,}
\keyword{frequence,}
