% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMeanPeriodLength.R
\name{findMeanPeriodLength}
\alias{findMeanPeriodLength}
\title{Finds the mean length of wet periods}
\usage{
findMeanPeriodLength(data, threshold = 0.1, d = F)
}
\arguments{
\item{data}{Data to do computations on, 3d array.}

\item{threshold}{Threshold for seting values to zero. Defaults to 0.1.}

\item{d}{If true, we calculate for drydays, else wet. Defaults to False}
}
\description{
This function findes the mean length of the wet periods for each gridpoint in the given dataset
}
\examples{
findMeanPeriodLength()
}
\keyword{dataHandling}
\keyword{length,}
\keyword{mean,}
\keyword{period,}
