% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeriodsLength.R
\name{findPeriodsLength}
\alias{findPeriodsLength}
\title{Finds length of wet periods.}
\usage{
findPeriodsLength(data, r, m, threshold = 0.1, d = F)
}
\arguments{
\item{data}{Data, 3d array, matrix, list or vector}

\item{r}{Row in matrix m to place data}

\item{m}{Matrix to place returned data.}

\item{threshold}{Threshold for seting values to zero. Defaults to 0.1.}

\item{d}{If true, we calculate for drydays, else wet. Defaults to False}
}
\description{
This function finds the length of wet periods in the dataset and places the data in a given matrix.
}
\examples{
findPeriodsLength()
}
\keyword{hist,}
\keyword{histogram,}
\keyword{length,}
\keyword{matrix}
\keyword{periods,}
