% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeriodsLengthGrid.R
\name{findPeriodsLengthGrid}
\alias{findPeriodsLengthGrid}
\title{Finds length of wet/dry periods per gridcell}
\usage{
findPeriodsLengthGrid(data, periodLength, frec = T, threshold = 0.1, d = F)
}
\arguments{
\item{data}{Data, 3d array}

\item{periodLength}{Length of one period/season in days}

\item{frec}{Return frequency instead of number of occurences. Defaults to TRUE}

\item{threshold}{Threshold for seting values to zero. Defaults to 0.1.}

\item{d}{If true, we calculate for drydays, else wet. Defaults to False}
}
\description{
This function finds the length of wet/dry periods in the dataset per gridcell, and return a matrix with the results (gridcell i, number of periods with length j).
}
\examples{
findPeriodsLengthGrid()
}
\keyword{grid}
\keyword{length,}
\keyword{matrix,}
\keyword{periods,}
