% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQD.R
\name{IQD}
\alias{IQD}
\title{Calculate the IQ distance at each location}
\usage{
IQD(modelArray, obsArray, crps, remove_indices = NULL)
}
\arguments{
\item{modelArray}{Array with model data, 3d-array (longitude, latitude, time)}

\item{obsArray}{Array with dataproduct data, 3d-array (longitude, latitude, time)}

\item{crps}{If true, crps is used (strongly recomended), otherwise integrate is used}

\item{remove_indices}{Which indices to remove. Defaults to NULL}
}
\description{
At each location, calculate IQ distance between the ecdf of a single data set 
of a climate model ('modelArray') and the ecdf of obervations ('obsArray').
Both input arrays are three-dimensional (longitude, latitude, time). The
function IQDlocal() is accessed with two options: the crps_sample function
(crps=TRUE) and the integrate funtion (crps=FALSE). The former is much faster
and with no problems during the calculations. The integrate function is
problematic due to the 'subdivisions' (the maximum number of subintervals),
i.e. no convergence even though the number of subintervals is large.
}
\examples{
IQD()
}
\keyword{IQD,}
\keyword{parallel}
