% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQDdiscreet.R
\name{IQDdiscreet}
\alias{IQDdiscreet}
\title{Calculate IQD between a and b}
\usage{
IQDdiscreet(X, Y, periodLength, dry = F, raw = F)
}
\arguments{
\item{periodLength}{Length of one period/season in days}

\item{dry}{If true, we calculate for drydays, else wet. Defaults to False}

\item{raw}{If true, input is expected to be raw model/dataproduct data, else result from findPeriodsLengthGrid. Defaults to False}

\item{a}{Data, 3d array}

\item{b}{Data, 3d array}
}
\description{
This function calculate the IQD between array a anb b and returns the answer as a grid, with IQD pr gridpoint
}
\examples{
IQDdiscreet()
}
\keyword{IQD,}
\keyword{discreet,}
\keyword{length,}
\keyword{matrix}
\keyword{periods,}
