% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQDlocal_crps.R
\name{IQDlocal_crps}
\alias{IQDlocal_crps}
\title{Calculate IQD with crps}
\usage{
IQDlocal_crps(modelVec, obsVec)
}
\arguments{
\item{modelVec}{Vector with model data}

\item{obsVec}{Vector with dataproduct data}
}
\description{
Calculate integrated quadratic distance (IQ distance) between
the empirical cumulative distribution functions (ecdfs) of
two data vectors with the crps_sample function.
}
\examples{
IQDlocal_crps()
}
\keyword{IQD,}
\keyword{IQDlocal,}
\keyword{crps}
