% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQDlocal_integrate.R
\name{IQDlocal_integrate}
\alias{IQDlocal_integrate}
\title{Calculate IQD with integration}
\usage{
IQDlocal_integrate(modelVec, obsVec)
}
\arguments{
\item{modelVec}{Vector with model data}

\item{obsVec}{Vector with dataproduct data}
}
\description{
Calculate integrated quadratic distance (IQ distance) between
the empirical cumulative distribution functions (ecdfs) of 
two data vectors with the integrate function. NB! The integrate
function is very slow and problematic with the 'subdivisions', 
(the maximum number of subintervals), i.e. no convergence even 
though the number of subintervals is large.
}
\examples{
IQDlocal_integrate()
}
\keyword{IQD,}
\keyword{IQDlocal,}
\keyword{integrate}
