% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMapPlot.R
\name{makeMapPlot}
\alias{makeMapPlot}
\title{Make a map plot}
\usage{
makeMapPlot(
  data,
  z_lab = "",
  option = "D",
  title = NULL,
  summ = NULL,
  info = NULL,
  title_col = "black",
  guide_lim = c(min(data$val), max(data$val)),
  log = FALSE,
  subtitle = NULL,
  ratio = 1.8,
  xlim = c(3, 35)
)
}
\arguments{
\item{data}{Data to plot}

\item{z_lab}{Defaults to ''}

\item{option}{Passed to "scale_colour_viridi" as option argument. Defaults to 'D'}

\item{title}{Passed to "labs" as title argument. Defaults to NULL,}

\item{summ}{Defaults to NULL}

\item{info}{Defaults to NULL}

\item{title_col}{Defaults to 'black'}

\item{guide_lim}{Defaults to c(min(data$val), max(data$val))}

\item{log}{If True, logscale. Defaults to FALSE}

\item{subtitle}{Passed to "labs" as subtitle argument. Defaults to NULL}

\item{ratio}{Defaults to 1.8}

\item{xlim}{xlimit. Defaults to c(3, 35)}
}
\description{
Used in severalGgPlots to make map plots
Takes in a data.frame with columns val, lon and lat.
Prints the values from the data.frame over a map of Fennoscandia
}
\examples{
makeMapPlot()
}
\keyword{plot,}
\keyword{severalGgPlots}
