% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMatrix.R
\name{makeMatrix}
\alias{makeMatrix}
\title{Make a list of pattern plots}
\usage{
makeMatrix(
  season,
  prepDataPath,
  depth = 2,
  seqVec = c("0-0-0", "0-0-1", "0-1-0", "0-1-1", "1-0-0", "1-0-1", "1-1-0", "1-1-1")
)
}
\arguments{
\item{season}{What season data to use.}

\item{prepDataPath}{Data path to prepered data.}

\item{depth}{Depth argument sent to psttree. Defaults to 1.}

\item{seqVec}{sequence vector sent to predict function. Defaults to c("0-0-0", "0-0-1", "0-1-0", "0-1-1", "1-0-0", "1-0-1", "1-1-0", "1-1-1").}
}
\description{
This function makes a list of pattern plots to be ploted in a matrix
}
\examples{
makeMatrix()
}
\keyword{list}
\keyword{matrix,}
\keyword{pattern,}
\keyword{patters,}
\keyword{plot,}
\keyword{plotlist,}
