% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScatterPlot.R
\name{makeScatterPlot}
\alias{makeScatterPlot}
\title{Creates a scatterplot}
\usage{
makeScatterPlot(plot_name_list, quantile_vals, phen, size = 3.5)
}
\arguments{
\item{plot_name_list}{Name of output plot with path per season}

\item{quantile_vals}{As created by severalGgPlots}

\item{phen}{Phenomenon, ex 'precipitation'}

\item{size}{Size of plots. Defaults to 3.5}
}
\description{
Creates a scatterplot for the phenomenon 'phen' for each season by using the list
'quantile_vals' created from function 'severalGgPlots'
plot_name_list contains names of plots per season
Also returns a list of the plots created
}
\examples{
makeScatterPlot()
}
\keyword{iqd,}
\keyword{plot,}
\keyword{plots}
\keyword{scatter,}
\keyword{scatterplot,}
