% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsNrCM.R
\name{ObsNrCM}
\alias{ObsNrCM}
\title{Find indexes for model}
\usage{
ObsNrCM(startEndEOBS, years)
}
\arguments{
\item{startEndEOBS}{}

\item{years}{Integer vector with value at index k year at index k from data array}
}
\description{
THIS SHOULD BE CORECCTED
This function findes which indexes to use for the model when comparing to the dataproduct, or somthing
}
\examples{
ObsNrCM()
}
\keyword{IQD}
