% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBarplot.R
\name{plotBarplot}
\alias{plotBarplot}
\title{Plots a map with image.plot using scico colorscales and map coast boundary}
\usage{
plotBarplot(
  data,
  title = "",
  season = "",
  frac = TRUE,
  rowNames = "",
  colList = c("#0000FF70", "#FF000070", "#00FF0070"),
  lines = F
)
}
\arguments{
\item{data}{Data to be ploted, 2d matrix or}

\item{title}{Title of plot. Defaults to "".}

\item{season}{Small text in upper corner of plot, for example to specify what season the data is from. Defaults to "".}

\item{frac}{True to plot fraction of periods, and not number of periods. Defaults to TRUE.}

\item{rowNames}{Names of rows, when plotting with matrix. Defaults to ".}

\item{colList}{List of colors supplied to plotfunction. Defaults to c("#0000FF70", "#FF000070", "#00FF0070").}

\item{lines}{If true plot lines insted of barplot. Defaults to False}
}
\description{
This function let you plot a map with drawn coast boundary, using a scico colorscale.
}
\examples{
plotBarplot()
}
\keyword{barplot,}
\keyword{histogram}
\keyword{plot,}
