% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plotMap}
\alias{plotMap}
\title{Plots a map with image.plot using scico colorscales and map coast boundary}
\usage{
plotMap(
  data,
  lon = lonFenno,
  lat = latFenno,
  title = "",
  season = "",
  seasonSize = 1,
  m = "",
  xlim = c(5, 32),
  ylim = c(55, 73),
  z = NULL,
  colScale = "roma",
  legend = T,
  lWidth = 4
)
}
\arguments{
\item{data}{Data to be ploted, 2d matrix}

\item{lon}{Longitude data, 2d matrix. Defaults to lonFenno.}

\item{lat}{Latitude data, 2d matrix. Defaults to latFenno.}

\item{title}{Title of plot. Defaults to "".}

\item{season}{Small text in upper corner of plot, for example to specify what season the data is from. Defaults to "".}

\item{seasonSize}{Text size of taxt passed to season argument. Defaults to 1.}

\item{m}{Supply mean to print in lower corner. Defaults to ""}

\item{xlim}{To limit map area in x direction. Defaults to [55, 32].}

\item{ylim}{To limit map area in y direction. Defaults to [5, 73].}

\item{z}{To set a scale for the colorbar. Defaults to ".}

\item{colScale}{What type of scico colorscale to use. Defaults to "roma".}

\item{legend}{If T plot legend using image.plot, else ploting using poly.image. Defaults to True}

\item{lWidth}{Width of legend. Defaults to 4}
}
\description{
This function let you plot a map with drawn coast boundary, using a scico colorscale.
}
\examples{
plotMap()
}
\keyword{map}
\keyword{maps,}
\keyword{plot,}
