% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPatternplot.R
\name{plotPatternplot}
\alias{plotPatternplot}
\title{Plot pattern plots}
\usage{
plotPatternplot(
  df,
  xlText = "",
  ylText = "",
  xticks = T,
  yticks = T,
  titleText = "",
  legend = T,
  showPanel = T,
  legendOnly = F,
  textSize = 4,
  n = 8,
  colScale = "grayC"
)
}
\arguments{
\item{df}{Pattern data to plot.}

\item{xlText}{Text to xlabel. Defaults to "".}

\item{ylText}{Text to ylabel. Defaults to "".}

\item{xticks}{True to display xticks. Defaults to T.}

\item{yticks}{True to display yticks. Defaults to T.}

\item{titleText}{Title of plot. Defaults to "".}

\item{legend}{True to display legend. Defaults to T.}

\item{showPanel}{True to keep plot frame. Defaults to T.}

\item{legendOnly}{True to only plot legend. Defaults to F.}

\item{textSize}{Size of text in plot. Defaults to 4.}

\item{n}{Number of colors, should equal number of possible combinations}

\item{colScale}{What type of scico colorscale to use. Defaults to "grayC".}
}
\description{
This function uses ggplot to plot pattern plots
}
\examples{
plotPatternplot()
}
\keyword{pattern,}
\keyword{patternplot}
\keyword{plot,}
