% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severalGgPlots.R
\name{severalGgPlots}
\alias{severalGgPlots}
\title{Creates a list of ggplots}
\usage{
severalGgPlots(
  file_paths_in,
  lon,
  lat,
  guide_lim,
  model_nrs = 1:length(model_names),
  quantiles = c(0.05, 0.95),
  want_summ = FALSE,
  method_nrs = 1:length(method_names),
  type_nrs = 1:length(type_names),
  prec = TRUE,
  log = FALSE,
  do_bootstrap = TRUE,
  bootstrap_samples = 1000
)
}
\arguments{
\item{file_paths_in}{file_paths for IQD data}

\item{lon}{Longitude coordinates}

\item{lat}{Latitude coordinates}

\item{guide_lim}{colour guide in plots}

\item{model_nrs}{Which models to plot. Defaults to 1:length(model_names)}

\item{quantiles}{Defaults to c(0.05, 0.95)}

\item{want_summ}{Defaults to FALSE}

\item{method_nrs}{Which methods to plot. Defaults to 1:length(method_names)}

\item{type_nrs}{Which dataproducts to plot. Defaults to 1:length(type_names)}

\item{prec}{If True, precipitation. Defaults to TRUE}

\item{log}{If True, logscale. Defaults to FALSE}

\item{do_bootstrap}{If True, use bootstrap. Defaults to TRUE}

\item{bootstrap_samples}{Number of bootstrap samples. Defaults to 1000}
}
\description{
Creates a list of ggplots for different combinations of dataset, season, model and method.
Input: file_paths for IQD data saved with each type of dataset,
       longitude and latitude matrices for all coordinates
       guide_limits used for the colour guide in the plots
Output: A list of ggplots.
        A list of mean and quantile values for each of the plots
}
\examples{
severalGgPlots()
}
\keyword{iqd,}
\keyword{plot,}
\keyword{plots}
\keyword{scatter,}
\keyword{scatterplot,}
